x-sheets
it is set-up and working for consultas.
looking forward and seeing what I have done until now, I realice that I
will have to customize a lot of the functionality and UI of
x-spreadsheets, and that it would actually require less work and would
be easier to just use plain HTML5 tables.